Add OpenMW project dependencies
There's a potential issue to figure out with this. The project page
shows libopenscenegraph-3.4-dev as a dependency, but this package can
no longer be found for Ubuntu 16.04. The libopenscenegraph-dev package
is available so I'll have to see if there are any issues trying to
build the project with this.